feat: adapt to LeviLamina 0.3.0

+++ src/Plugin.cpp
@@ -1,19 +1,20 @@
 #include "Plugin.h"
 
 #include <memory>
+#include <stdexcept>
 #include <string>
-#include <utility>
+
 
 #include <ll/api/Config.h>
 #include <ll/api/base/KeyValueDB.h>
 #include <ll/api/command/DynamicCommand.h>
 #include <ll/api/event/EventBus.h>
 #include <ll/api/event/ListenerBase.h>
-#include <ll/api/event/command/SetupCommandEvent.h>
 #include <ll/api/event/player/PlayerJoinEvent.h>
 #include <ll/api/event/player/PlayerUseItemEvent.h>
 #include <ll/api/form/ModalForm.h>
 #include <ll/api/plugin/NativePlugin.h>
+#include <ll/api/service/Bedrock.h>
 #include <mc/entity/utilities/ActorType.h>
 #include <mc/server/commands/CommandOrigin.h>
 #include <mc/server/commands/CommandOutput.h>
@@ -21,12 +22,11 @@
 #include <mc/world/actor/player/Player.h>
 #include <mc/world/item/registry/ItemStack.h>
 
+
 namespace plugin {
 
 Plugin::Plugin(ll::plugin::NativePlugin& self) : mSelf(self) {
     auto& logger = mSelf.getLogger();
-
-    logger.info("Loading...");
 
     // Load or initialize configurations.
     const auto& configFilePath = self.getConfigDir() / "config.json";
@@ -42,14 +42,10 @@ Plugin::Plugin(ll::plugin::NativePlugin& self) : mSelf(self) {
     // Initialize databases;
     const auto& playerDbPath = self.getDataDir() / "players";
     mPlayerDb                = std::make_unique<ll::KeyValueDB>(playerDbPath);
-
-    logger.info("Loaded");
 }
 
 bool Plugin::enable() {
     auto& logger = mSelf.getLogger();
-
-    logger.info("Enabling...");
 
     // Subscribe to events.
     auto& eventBus = ll::event::EventBus::getInstance();
@@ -96,7 +92,7 @@ bool Plugin::enable() {
                         "Yes",
                         "No",
                         [&logger](Player& player, bool yes) {
-                            if (!yes) {
+                            if (yes) {
                                 player.kill();
 
                                 logger.info("{} killed themselves", player.getRealName());
@@ -109,30 +105,30 @@ bool Plugin::enable() {
             }
         });
 
-    mSetupCommandEventListener = eventBus.emplaceListener<
-        ll::event::command::SetupCommandEvent>([&logger](ll::event::command::SetupCommandEvent& event) {
-        // Setup suicide command.
-        auto command =
-            DynamicCommand::createCommand(event.registry(), "suicide", "Commits suicide.", CommandPermissionLevel::Any);
-        command->addOverload();
-        command->setCallback(
-            [&logger](DynamicCommand const&, CommandOrigin const& origin, CommandOutput& output, std::unordered_map<std::string, DynamicCommand::Result>&) {
-                auto* entity = origin.getEntity();
-                if (entity == nullptr || !entity->isType(ActorType::Player)) {
-                    output.error("Only players can commit suicide");
-                    return;
-                }
+    // Register commands.
+    auto commandRegistry = ll::service::getCommandRegistry();
+    if (!commandRegistry) {
+        throw std::runtime_error("failed to get command registry");
+    }
 
-                auto* player = static_cast<Player*>(entity);
-                player->kill();
-
-                logger.info("{} killed themselves", player->getRealName());
+    auto command =
+        DynamicCommand::createCommand(commandRegistry, "suicide", "Commits suicide.", CommandPermissionLevel::Any);
+    command->addOverload();
+    command->setCallback(
+        [&logger](DynamicCommand const&, CommandOrigin const& origin, CommandOutput& output, std::unordered_map<std::string, DynamicCommand::Result>&) {
+            auto* entity = origin.getEntity();
+            if (entity == nullptr || !entity->isType(ActorType::Player)) {
+                output.error("Only players can commit suicide");
+                return;
             }
-        );
-        DynamicCommand::setup(event.registry(), std::move(command));
-    });
 
-    logger.info("Enabled");
+            auto* player = static_cast<Player*>(entity);
+            player->kill();
+
+            logger.info("{} killed themselves", player->getRealName());
+        }
+    );
+    DynamicCommand::setup(commandRegistry, std::move(command));
 
     return true;
 }
@@ -140,15 +136,18 @@ bool Plugin::enable() {
 bool Plugin::disable() {
     auto& logger = mSelf.getLogger();
 
-    logger.info("Disabling...");
-
     // Unsubscribe from events.
     auto& eventBus = ll::event::EventBus::getInstance();
     eventBus.removeListener(mPlayerJoinEventListener);
     eventBus.removeListener(mPlayerUseItemEventListener);
-    eventBus.removeListener(mSetupCommandEventListener);
 
-    logger.info("Disabled");
+    // Unregister commands.
+    auto commandRegistry = ll::service::getCommandRegistry();
+    if (!commandRegistry) {
+        throw std::runtime_error("failed to get command registry");
+    }
+
+    commandRegistry->unregisterCommand("suicide");
 
     return true;
 }
